Moving API event handler code to cluster package
