+++ api/primary.go
@@ -14,7 +14,7 @@ import (
 // Primary router context, used by handlers.
 type context struct {
 	cluster       cluster.Cluster
-	eventsHandler *eventsHandler
+	eventsHandler *cluster.EventsHandler
 	listenerCount *uint64
 	statusHandler StatusHandler
 	debug         bool
@@ -123,7 +123,7 @@ func NewPrimary(cluster cluster.Cluster, tlsConfig *tls.Config, status StatusHan
 	// is initialized with no writers or channels. This is in api/events.go and
 	// uses the watch package from SwarmKit, which is based on the go-events
 	// package. See https://github.com/docker/swarm/issues/2718 for context
-	eventsHandler := newEventsHandler()
+	eventsHandler := cluster.NewAPIEventsHandler()
 	listenerCount := uint64(0)
 	// need to add this queue to the cluster
 	// This just calls c.eventHandlers.RegisterEventHandler(eventsHandler) internally.
